Add support for com.google.common.collect.Table
The serialized form of tables mirrors the structure of table.rowMap().
For example, a table containing these values:

     a  b  c
  x  1  2  3
  y  4  5  6

should have a JSON form of:

{
  "x": { "a": 1, "b": 2, "c": 3 },
  "y": { "a": 4, "b": 5, "c": 6 }
}

Any table can be serialized.  Only four table types can be
deserialized:

 1. Table - uses an ImmutableTable
 2. ImmutableTable
 3. HashBasedTable
 4. ArrayTable

It is not necessarily the case that the serialized form of one table
type can be deserialized into another table type.  ArrayTable may
contain null values, which the other table types do not support.  (And
the null values cannot be omitted from a serialized ArrayTable if we
expect the deserialized ArrayTable to be equal.  The universe of row
and column keys in an ArrayTable is fixed.  If null values were
omitted during serialization, the knowledge of which row and column
keys are supposed to exist might be lost.)

All serialization and deserialization is accomplished through
StdDelegatingSerializer and StdDelegatingDeserializer.  Tables are
converted to and from Map instances before they are converted to or
from JSON:

  Table -> Map -> JSON -> Map -> Table

A more direct, non-delegating implentation might have been more
efficient, but this approach was simpler.

Deserialization uses LinkedHashMap for the Map implementation to
preserve cell ordering, in case the Table implementation also makes
guarantees about ordering.  (Only HashBasedTable does not.)